refactor: change dropdown to named export
